refactor(client): git hooks to git folder
